Custom_Functions/requirements.txt
faiss-cpu==1.8.0
langchain==0.2.12
langchain-openai==0.1.22
langchain-community==0.2.11
openai==1.35.10
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.36.0